% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geopackage_data.R
\name{get_geopackage_data}
\alias{get_geopackage_data}
\title{Retrieves geopackage integration from GADM (https://gadm.org)}
\usage{
get_geopackage_data(country_code, file_layer)
}
\arguments{
\item{country_code}{String. The ISO country code for the desired country.}

\item{file_layer}{Integer. The layer of the geopackage to access. Defaults to last available layer.}
}
\value{
The geopackage for the specified country and layer.
}
\description{
Retrieves geopackage integration from GADM (https://gadm.org)
}
\examples{
# Retrieve geopackage data for Spain (ESP) and layer 4
get_geopackage_data(country_code = "ESP", file_layer = 4)
}
