% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_aggregates.R
\name{get_malert_aggregates}
\alias{get_malert_aggregates}
\title{Aggregates mosquito alert report data by country or city}
\usage{
get_malert_aggregates(
  aggregate_type,
  filter_year = NULL,
  country_code = NULL,
  file_layer = NULL,
  ma_filters = NULL
)
}
\arguments{
\item{aggregate_type}{String. The type of aggregation to perform. Options are "country" or "city".}

\item{filter_year}{String. The year(s) to filter the data. Options include a single year (e.g., "2014"),
a comma-separated list of years (e.g., "2014,2015"), or a range (e.g., "2014-2016"). Defaults to NULL (all years).}

\item{country_code}{String. The ISO country code (required if aggregate_type is "city").}

\item{file_layer}{Integer. The layer of the shapefile/geopackage to access (for city aggregation).}

\item{ma_filters}{Optional named list of filters to apply to Mosquito Alert data.
Each list element name should be a column name and values can be scalars or vectors, e.g.
\code{list("movelab_annotation_euro.class_id" = c(4, 5), type = "adult")}.}
}
\value{
A data frame containing the aggregated mosquito alert report counts.
}
\description{
Aggregates mosquito alert report data by country or city
}
\examples{
# Aggregate mosquito reports by country for 2015, 2016 and 2018
get_malert_aggregates(aggregate_type = "country", filter_year = "2015,2016,2018")

# Aggregate mosquito reports by city for Spain in 2014-2024
get_malert_aggregates(aggregate_type = "city", filter_year = "2014-2024", country_code = "ESP", file_layer = 2)
}
