% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_species_counts.R
\name{get_malert_species_counts}
\alias{get_malert_species_counts}
\title{Count Mosquito Alert reports by species}
\usage{
get_malert_species_counts(
  source = "zenodo",
  doi = "10.5281/zenodo.597466",
  iso3 = NULL,
  admin_level = NULL,
  admin_name = NULL
)
}
\arguments{
\item{source}{Source to download from: \code{"github"} or \code{"zenodo"}.}

\item{doi}{Zenodo DOI. Defaults to the DOI for the latest version.}

\item{iso3}{Optional three-letter ISO country code for spatial filtering.}

\item{admin_level}{Optional administrative level for spatial filtering.}

\item{admin_name}{Optional administrative name for spatial filtering.}
}
\value{
A tibble containing
\code{movelab_annotation_euro.class_name} and \code{count}.
}
\description{
Downloads Mosquito Alert report data using \code{\link[=get_malert_data]{get_malert_data()}} and returns
the number of reports for each species.
}
