% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_geopackage_data.R
\name{get_malert_geopackage_data}
\alias{get_malert_geopackage_data}
\title{Generates mosquito alert reports with geopackage}
\usage{
get_malert_geopackage_data(filter_year, country_code, file_layer)
}
\arguments{
\item{filter_year}{String. The year(s) you would like to filter for. Default is all. "2014,2"}

\item{country_code}{String. Country you would like to select}

\item{file_layer}{String. the layer of the geopackage to access}
}
\value{
The aggregated data.
}
\description{
Generates mosquito alert reports with geopackage
}
\examples{
malert_reports = get_malert_data(source = "github")
malert_reports
}
